Splits sauce tests into test-sauce task
- `make test` now only runs tests PhantomJS
- `make test-sauce` runs tests in PhantomJS and saucelabs
- travis runs `make test-sauce` by default
